updated dependencies and made gold noise highp to fix render artifacts on iPad.

+++ src/shaders/shaders.vr.fragment.js
@@ -49,11 +49,11 @@ export default class ShadersFragment {
 
 // Gold Noise ©2015 dcerisano@standard3d.com 
 // https://stackoverflow.com/questions/4200224/random-noise-functions-for-glsl
-const lowp float PHI = 1.61803398874989484820459 * 00000.1; // Golden Ratio   
-const lowp float PI  = 3.14159265358979323846264 * 00000.1; // PI
-const lowp float SQ2 = 1.41421356237309504880169 * 10000.0; // Square Root of Two
+const highp float PHI = 1.61803398874989484820459 * 00000.1; // Golden Ratio   
+const highp float PI  = 3.14159265358979323846264 * 00000.1; // PI
+const highp float SQ2 = 1.41421356237309504880169 * 10000.0; // Square Root of Two
 
-lowp float gold_noise(in lowp vec2 coordinate, in lowp float seed){
+highp float gold_noise(in highp vec2 coordinate, in highp float seed){
     return fract(tan(distance(coordinate*(seed+PHI), vec2(PHI, PI)))*SQ2);
 }
